+++ tests/__lib/test-context/git.ts
@@ -1,13 +1,43 @@
+import * as nodefs from 'fs'
 import isogit from 'isomorphic-git'
-import { createGit } from '../../../src/lib/git'
 import { TmpDirContribution } from './tmp-dir'
 
-interface GitContribution {
-  git: typeof isogit
-}
+const fs = nodefs
 
-export const git = (ctx: TmpDirContribution): GitContribution => {
-  const git = createGit()
-  git.cwd(ctx.dir)
-  return { git: isogit }
+export const git = (ctx: TmpDirContribution) => {
+  const dir = ctx.dir
+  const git = isogit
+
+  return {
+    git: isogit,
+    commit(message: string) {
+      return git.commit({ fs, dir, message, author: { name: 'labs' } })
+    },
+    // https://github.com/isomorphic-git/isomorphic-git/issues/129#issuecomment-390884874
+    async hardReset({ dir, ref, branch }: { dir: string; ref: string; branch: string }) {
+      var re = /^HEAD~([0-9]+)$/i
+      var m = ref.match(re)
+      if (m) {
+        var count = +m[1]
+        var commits = await git.log({ fs, dir, depth: count + 1 })
+        var commit = commits.pop()!.oid
+        return new Promise((resolve, reject) => {
+          fs.writeFile(dir + `/.git/refs/heads/${branch}`, commit, (err) => {
+            if (err) {
+              return reject(err)
+            }
+            // clear the index (if any)
+            fs.unlink(dir + '/.git/index', (err) => {
+              if (err) {
+                return reject(err)
+              }
+              // checkout the branch into the working tree
+              git.checkout({ dir, fs, ref: branch }).then(resolve)
+            })
+          })
+        })
+      }
+      return Promise.reject(`Wrong ref ${ref}`)
+    },
+  }
 }
